+++ docs/news.rst
@@ -7,6 +7,17 @@
 ``psycopg`` release notes
 =========================
 
+Future releases
+---------------
+
+Psycopg 3.1.13 (unreleased)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+- Raise `DataError` instead of whatever internal failure trying to dump a
+  `~datetime.time` object with with a `!tzinfo` specified as
+  `~zoneinfo.ZoneInfo` (ambiguous offset, see :ticket:`#652`).
+
+
 Current release
 ---------------